When a diff is baked, the generated migration file should be marked as migrated
