+++ src/Shell/Task/MigrationDiffTask.php
@@ -443,7 +443,6 @@ class MigrationDiffTask extends SimpleMigrationTask
 
         $input = new ArrayInput($inputArgs, $definition);
         $path = $this->getOperationsPath($input) . DS . 'schema-dump-' . $connectionName . '.lock';
-        debug($path);
 
         if (!file_exists($path)) {
             $msg = 'Unable to retrieve the schema dump file. You can create a dump file using ' .
