+++ src/Shell/Task/SnapshotTrait.php
@@ -39,6 +39,50 @@ trait SnapshotTrait
     public $skipTablesRegex = '_phinxlog';
 
     /**
+     * After a file has been successfully created, we mark the newly
+     * created migration as applied
+     *
+     * @param string $path Where to put the file.
+     * @param string $contents Content to put in the file.
+     * @return bool Success
+     */
+    public function createFile($path, $contents)
+    {
+        $createFile = parent::createFile($path, $contents);
+
+        if ($createFile) {
+            $this->markSnapshotApplied($path);
+        }
+
+        return $createFile;
+    }
+
+    /**
+     * Will mark a snapshot created, the snapshot being identified by its
+     * full file path.
+     *
+     * @param string $path Path to the newly created snapshot
+     * @return void
+     */
+    protected function markSnapshotApplied($path)
+    {
+        $fileName = pathinfo($path, PATHINFO_FILENAME);
+        list($version, ) = explode('_', $fileName, 2);
+
+        $dispatchCommand = 'migrations mark_migrated -t ' . $version . ' -o';
+        if (!empty($this->params['connection'])) {
+            $dispatchCommand .= ' -c ' . $this->params['connection'];
+        }
+
+        if (!empty($this->params['plugin'])) {
+            $dispatchCommand .= ' -p ' . $this->params['plugin'];
+        }
+
+        $this->_io->out('Marking the migration ' . $fileName . ' as migrated...');
+        $this->dispatchShell($dispatchCommand);
+    }
+
+    /**
      * Gets a list of table to baked based on the Collection instance passed and the options passed to
      * the shell call.
      *
